refactor: separate files for different classes
